+++ requests_cache/policy/settings.py
@@ -1,4 +1,4 @@
-from typing import Dict, Iterable, Union, Optional
+from typing import Dict, Iterable, Optional, Union
 
 from attrs import define, field
 
@@ -36,6 +36,7 @@ class CacheSettings(RichMixin):
     key_fn: KeyCallback = field(default=None)
     match_headers: Union[Iterable[str], bool] = field(default=False)
     only_if_cached: bool = field(default=False)
+    read_only: bool = field(default=False)
     stale_if_error: Union[bool, ExpirationTime] = field(default=False)
     stale_while_revalidate: Union[bool, ExpirationTime] = field(default=False)
     urls_expire_after: Dict[ExpirationPattern, ExpirationTime] = field(factory=dict)